Refactor text_to_cstring utility: move to utils.hpp and implement in utils.cpp
