+++ src/utils.cpp
@@ -8,11 +8,29 @@
 #include <string>
 #include <vector>
 
+extern "C" {
+#include <utils/builtins.h>
+}
+
 #include <nlohmann/json.hpp>
 
 #include "./include/logger.hpp"
 
 namespace pg_ai::utils {
+
+std::string pg_text_to_string(const text* t) {
+  if (!t)
+    return "";
+  char* cstr = text_to_cstring(t);
+  try {
+    std::string result(cstr);
+    pfree(cstr);
+    return result;
+  } catch (...) {
+    pfree(cstr);
+    throw;
+  }
+}
 
 std::pair<bool, std::string> read_file(const std::string& filepath) {
   std::ifstream file(filepath, std::ios::binary | std::ios::ate);
